bin/scripts/changelog/requirements-changelog.txt
# Changelog Generator Dependencies
GitPython>=3.1.40
sqlparse>=0.4.4
click>=8.1.7
python-dateutil>=2.8.2
PyYAML>=6.0.1
